feat: refine tests to special services
